requirements.txt
fastapi
uvicorn
pydantic
email-validator
python-jose[cryptography]
python-multipart